Use a config to generate the layer import dialog
The LayerImport Dialog can differ with different geonode instances,
this adds the ability to specify what is displayed in the importer
window
There are a few preconfigured elements, that can be turned on/off
for example the layerName and permssion handling.
The config will be set via reduy and read from there eventually.
As well as the saving for all the fields in the other section